feat(Initial Commit): This is the inital commit for the Project
In this commit i have made a new folder for the frontend and created NavBar , Drawer And a layout for the Planner.

I have used Framer motion for smooth animation which uses GPU acceleration for animation purposes.

I have used beercss for Styling purposes and better Developer Experience.

React for rendering purpose.